Renamed AddDialog class to CreateBagDialog and removed possibility of creating a bag after it has already been created
