+++ Swing-Lab-Work/src/gui/CreateBagDialog.form
@@ -1,12 +1,57 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="gui.CreateBagDialog">
-  <grid id="27dc6" binding="panel" layout-manager="GridLayoutManager" row-count="1" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
+  <grid id="27dc6" binding="panel" layout-manager="GridLayoutManager" row-count="3" column-count="5" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
-      <xy x="20" y="20" width="302" height="195"/>
+      <xy x="20" y="20" width="502" height="393"/>
     </constraints>
-    <properties/>
+    <properties>
+      <background color="-14737632"/>
+      <foreground color="-15921907"/>
+    </properties>
     <border type="none"/>
-    <children/>
+    <children>
+      <vspacer id="688cd">
+        <constraints>
+          <grid row="1" column="1" row-span="2" col-span="1" vsize-policy="6" hsize-policy="1" anchor="8" fill="2" indent="0" use-parent-layout="false"/>
+        </constraints>
+      </vspacer>
+      <component id="1c18a" class="javax.swing.JLabel">
+        <constraints>
+          <grid row="0" column="1" row-span="1" col-span="3" vsize-policy="0" hsize-policy="0" anchor="0" fill="0" indent="0" use-parent-layout="false">
+            <preferred-size width="150" height="100"/>
+          </grid>
+        </constraints>
+        <properties>
+          <font name="Comic Sans MS" size="48" style="1"/>
+          <foreground color="-526345"/>
+          <text value="Enter a Bag Volume"/>
+        </properties>
+      </component>
+      <component id="a6d50" class="javax.swing.JTextField" binding="textField">
+        <constraints>
+          <grid row="1" column="2" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="0" fill="0" indent="0" use-parent-layout="false">
+            <preferred-size width="200" height="50"/>
+          </grid>
+        </constraints>
+        <properties>
+          <doubleBuffered value="false"/>
+          <editable value="true"/>
+          <enabled value="true"/>
+          <font name="Comic Sans MS" size="48" style="1"/>
+          <text value=""/>
+          <verifyInputWhenFocusTarget value="true"/>
+        </properties>
+      </component>
+      <component id="b9564" class="javax.swing.JButton" binding="OKButton" default-binding="true">
+        <constraints>
+          <grid row="2" column="2" row-span="1" col-span="2" vsize-policy="0" hsize-policy="3" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <font name="Comic Sans MS" size="36" style="1"/>
+          <text value="OK"/>
+        </properties>
+      </component>
+    </children>
   </grid>
 </form>
