+++ Swing-Lab-Work/src/gui/CreateBagDialog.java
@@ -2,10 +2,30 @@ package gui;
 
 import javax.swing.*;
 import java.awt.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
 
 public class CreateBagDialog extends JDialog {
 
+    private JTextField textField;
+    private JButton OKButton;
     private JPanel panel;
+
+    CreateBagDialog(MainFrame mainFrame) {
+        super();
+        setContentPane(panel);
+        setLocation(600, 300);
+        setSize(500, 300);
+
+        OKButton.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                mainFrame.getBag().setVolume(Double.parseDouble(textField.getText()));
+                dispose();
+                mainFrame.update();
+            }
+        });
+    }
 
     {
 // GUI initializer generated by IntelliJ IDEA GUI Designer
@@ -23,7 +43,50 @@ public class CreateBagDialog extends JDialog {
      */
     private void $$$setupUI$$$() {
         panel = new JPanel();
-        panel.setLayout(new com.intellij.uiDesigner.core.GridLayoutManager(1, 1, new Insets(0, 0, 0, 0), -1, -1));
+        panel.setLayout(new com.intellij.uiDesigner.core.GridLayoutManager(3, 5, new Insets(0, 0, 0, 0), -1, -1));
+        panel.setBackground(new Color(-14737632));
+        panel.setForeground(new Color(-15921907));
+        final com.intellij.uiDesigner.core.Spacer spacer1 = new com.intellij.uiDesigner.core.Spacer();
+        panel.add(spacer1, new com.intellij.uiDesigner.core.GridConstraints(1, 1, 2, 1, com.intellij.uiDesigner.core.GridConstraints.ANCHOR_WEST, com.intellij.uiDesigner.core.GridConstraints.FILL_VERTICAL, 1, com.intellij.uiDesigner.core.GridConstraints.SIZEPOLICY_WANT_GROW, null, null, null, 0, false));
+        final JLabel label1 = new JLabel();
+        Font label1Font = this.$$$getFont$$$("Comic Sans MS", Font.BOLD, 48, label1.getFont());
+        if (label1Font != null) label1.setFont(label1Font);
+        label1.setForeground(new Color(-526345));
+        label1.setText("Enter a Bag Volume");
+        panel.add(label1, new com.intellij.uiDesigner.core.GridConstraints(0, 1, 1, 3, com.intellij.uiDesigner.core.GridConstraints.ANCHOR_CENTER, com.intellij.uiDesigner.core.GridConstraints.FILL_NONE, com.intellij.uiDesigner.core.GridConstraints.SIZEPOLICY_FIXED, com.intellij.uiDesigner.core.GridConstraints.SIZEPOLICY_FIXED, null, new Dimension(150, 100), null, 0, false));
+        textField = new JTextField();
+        textField.setDoubleBuffered(false);
+        textField.setEditable(true);
+        textField.setEnabled(true);
+        Font textFieldFont = this.$$$getFont$$$("Comic Sans MS", Font.BOLD, 48, textField.getFont());
+        if (textFieldFont != null) textField.setFont(textFieldFont);
+        textField.setText("");
+        textField.setVerifyInputWhenFocusTarget(true);
+        panel.add(textField, new com.intellij.uiDesigner.core.GridConstraints(1, 2, 1, 1, com.intellij.uiDesigner.core.GridConstraints.ANCHOR_CENTER, com.intellij.uiDesigner.core.GridConstraints.FILL_NONE, com.intellij.uiDesigner.core.GridConstraints.SIZEPOLICY_WANT_GROW, com.intellij.uiDesigner.core.GridConstraints.SIZEPOLICY_FIXED, null, new Dimension(200, 50), null, 0, false));
+        OKButton = new JButton();
+        Font OKButtonFont = this.$$$getFont$$$("Comic Sans MS", Font.BOLD, 36, OKButton.getFont());
+        if (OKButtonFont != null) OKButton.setFont(OKButtonFont);
+        OKButton.setText("OK");
+        panel.add(OKButton, new com.intellij.uiDesigner.core.GridConstraints(2, 2, 1, 2, com.intellij.uiDesigner.core.GridConstraints.ANCHOR_CENTER, com.intellij.uiDesigner.core.GridConstraints.FILL_NONE, com.intellij.uiDesigner.core.GridConstraints.SIZEPOLICY_CAN_SHRINK | com.intellij.uiDesigner.core.GridConstraints.SIZEPOLICY_CAN_GROW, com.intellij.uiDesigner.core.GridConstraints.SIZEPOLICY_FIXED, null, null, null, 0, false));
+    }
+
+    /**
+     * @noinspection ALL
+     */
+    private Font $$$getFont$$$(String fontName, int style, int size, Font currentFont) {
+        if (currentFont == null) return null;
+        String resultName;
+        if (fontName == null) {
+            resultName = currentFont.getName();
+        } else {
+            Font testFont = new Font(fontName, Font.PLAIN, 10);
+            if (testFont.canDisplay('a') && testFont.canDisplay('1')) {
+                resultName = fontName;
+            } else {
+                resultName = currentFont.getName();
+            }
+        }
+        return new Font(resultName, style >= 0 ? style : currentFont.getStyle(), size >= 0 ? size : currentFont.getSize());
     }
 
     /**
@@ -32,4 +95,5 @@ public class CreateBagDialog extends JDialog {
     public JComponent $$$getRootComponent$$$() {
         return panel;
     }
+
 }
